Former-commit-id: be8cef1df5b8d56e7af3cf98b5739fe6befcdaee

+++ Sources/Internal/Entity/Component.h
@@ -137,7 +137,7 @@ inline Entity* Component::GetEntity() const
     virtual uint32 GetType() const { return TYPE; }; \
     static const uint32 C_TYPE = TYPE; 
 
-#define MAKE_COMPONENT_MASK(x) ((uint64)1 << (uint64)x)
+#define MAKE_COMPONENT_MASK(x) (1ULL << static_cast<uint64>(x))
 
 template <template <typename> class Container, class T>
 void Component::GetDataNodes(Container<T>& container)
